sql: fix potential OoB access in lemon.c
The static analyzer found a potential OoB problem in the lemon.c file.
This situation occurs when you go beyond the boundaries of the
array x in the Rule_sort() function.

To avoid this problem, we increased length of the array x.

NO_DOC=only theoretical vulnerability
NO_CHANGELOG=only theoretical vulnerability
NO_TEST=currently it is impossible to create a test