lib: introduce data bank
Data bank is a small library that is designed for simplification
of allocation of several objects in one memory block. It could be
anything, but special attention is given to string objects, that
are arrays of chars.

Typical usage consist of two phases: gathering total needed size
of memory block and creation of objects in given block.

NO_DOC=refactoring
NO_CHANGELOG=refactoring