box: add pin/unping infrastructure for spaces
There are cases when we need to be sure that a space by given
id or name is not deleted; and if it is replaced (in space cache),
there's need to track pointer to new space. Like ib constraints:
they must hold a pointer to struct space while it's very hard to
determine whether there'a constraint that points to given space.

Implement space pin/unpin for this purpose. You can pin a space to
declare that the space is require to exist. To have to unpin it
when the space is not needed anymore.

NO_DOC=refactoring
NO_CHANGELOG=refactoring